free.py, visualize ints instead of floats

+++ examples/free.py
@@ -18,15 +18,15 @@ def main():
     swap = psutil.swap_memory()
     templ = "%-7s %10s %10s %10s %10s %10s %10s"
     print_(templ % ('', 'total', 'used', 'free', 'shared', 'buffers', 'cache'))
-    print_(templ % ('Mem:', virt.total / 1024,
-                            virt.used / 1024,
-                            virt.free / 1024,
-                            getattr(virt, 'shared', 0) / 1024,
-                            getattr(virt, 'buffers', 0) / 1024,
-                            getattr(virt, 'cached', 0) / 1024))
-    print_(templ % ('Swap:', swap.total / 1024,
-                             swap.used / 1024,
-                             swap.free / 1024,
+    print_(templ % ('Mem:', int(virt.total / 1024),
+                            int(virt.used / 1024),
+                            int(virt.free / 1024),
+                            int(getattr(virt, 'shared', 0) / 1024),
+                            int(getattr(virt, 'buffers', 0) / 1024),
+                            int(getattr(virt, 'cached', 0) / 1024)))
+    print_(templ % ('Swap:', int(swap.total / 1024),
+                             int(swap.used / 1024),
+                             int(swap.free / 1024),
                              '', '', ''))
 
 if __name__ == '__main__':
